requirements.txt
autogen-agentchat @ git+https://github.com/patrickstolc/autogen.git@0.2#egg=autogen-agentchat # AutoGen AgentChat extension (Forked)
autogen==0.3.1 # Core AutoGen library
google-generativeai # Required for Google Gemini API client
fix-busted-json==0.0.18 # For repairing malformed JSON responses
python-dotenv==1.0.1  # For loading environment variables from .env files

requests  # For making HTTP requests
pandas  # For data manipulation and analysis